Added functionality to test if the foreground and background images have been segmented.
This is a full stack change requiring new functions to be written in most places.  We add a
callback function to the API calls load_foreground and load_background, and an Observer pattern
object instance to the TF_server object to act as an arbiter between the segmentation functions
and the call asking if an image has been segmented.  This Observer class uses locking to attempt
safe concurrent access.